+++ extensions/vscode/src/views/homeView.ts
@@ -97,8 +97,6 @@ enum HomeViewInitialized {
 export class HomeViewProvider implements WebviewViewProvider, Disposable {
   private disposables: Disposable[] = [];
 
-  private state: PublisherState;
-
   private root: WorkspaceFolder | undefined;
   private webviewView?: WebviewView;
   private extensionUri: Uri;
@@ -118,13 +116,12 @@ export class HomeViewProvider implements WebviewViewProvider, Disposable {
   constructor(
     private readonly context: ExtensionContext,
     private readonly stream: EventStream,
+    private state: PublisherState,
   ) {
     const workspaceFolders = workspace.workspaceFolders;
     if (workspaceFolders !== undefined) {
       this.root = workspaceFolders[0];
     }
-
-    this.state = new PublisherState(this.context);
 
     this.extensionUri = this.context.extensionUri;
     this.webviewConduit = new WebviewConduit();
@@ -233,9 +230,9 @@ export class HomeViewProvider implements WebviewViewProvider, Disposable {
         configurationName,
         !extensionSettings.verifyCertificates(), // insecure = !verifyCertificates
         projectDir,
-        secrets,
         r,
         python,
+        secrets,
       );
       deployProject(
         deploymentName,
@@ -278,6 +275,13 @@ export class HomeViewProvider implements WebviewViewProvider, Disposable {
     // On first run, we have no saved state. Trigger a save
     // so we have the state, and can notify dependent views.
     this.requestWebviewSaveSelection();
+
+    // Listen for credential updates.
+    this.disposables.push(
+      this.state.onDidRefreshCredentials((_e) => {
+        this.updateWebViewViewCredentials();
+      }),
+    );
 
     // Signal the webapp that we believe the initialization refreshes
     // are finished.
@@ -383,7 +387,6 @@ export class HomeViewProvider implements WebviewViewProvider, Disposable {
 
   private async refreshCredentials() {
     await this.state.refreshCredentials();
-    return this.updateWebViewViewCredentials();
   }
 
   private async refreshActiveConfig(cfg?: Configuration | ConfigurationError) {
@@ -782,8 +785,8 @@ export class HomeViewProvider implements WebviewViewProvider, Disposable {
 
           return await api.packages.createRRequirementsFile(
             activeConfiguration.projectDir,
-            relPathPackageFile,
             r,
+            relPathPackageFile,
           );
         },
       );