+++ extensions/vscode/CHANGELOG.md
@@ -6,6 +6,13 @@ file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Credentials are now exposed via a `posit-connect` authentication provider,
+  which surfaces them in the native Accounts UI. (#2608)
+
 ## [1.10.0]
 
 ### Added